+++ src/include/storage/storage_util.h
@@ -128,7 +128,7 @@ class StorageUtil {
    *
    * @return {offset_varlen, offset_8, offset_4, offset_2, offset_1}
    */
-  static std::vector<uint16_t> ComputeBaseAttributeOffsets(const std::vector<uint8_t> &attr_sizes,
+  static std::vector<uint16_t> ComputeBaseAttributeOffsets(const std::vector<uint16_t> &attr_sizes,
                                                            uint16_t num_reserved_columns);
 
   /**
